+++ components/Icon/index.js
@@ -1,5 +1,14 @@
 const Icon = ({ children: icon, className, white }) => (
-  <>{icon ? <img className={className} title={icon} src={require(`./${icon.toLowerCase() + (white ? 'White' : '')}.svg`)} alt={icon} /> : null}</>
+  <>
+    {icon ? (
+      <img
+        className={className}
+        title={icon}
+        src={require(`./${icon.toLowerCase() + (white ? 'White' : '')}.svg`)}
+        alt={icon}
+      />
+    ) : null}
+  </>
 )
 
-export default Icon;
+export default Icon
